Move DestinationTableBuilder to BigQueryClient
