+++ plugin/trino-bigquery/src/main/java/io/trino/plugin/bigquery/ReadSessionCreator.java
@@ -13,12 +13,6 @@
  */
 package io.trino.plugin.bigquery;
 
-import com.google.cloud.BaseServiceException;
-import com.google.cloud.bigquery.BigQueryException;
-import com.google.cloud.bigquery.Job;
-import com.google.cloud.bigquery.JobInfo;
-import com.google.cloud.bigquery.QueryJobConfiguration;
-import com.google.cloud.bigquery.Table;
 import com.google.cloud.bigquery.TableDefinition;
 import com.google.cloud.bigquery.TableId;
 import com.google.cloud.bigquery.TableInfo;
@@ -33,15 +27,12 @@ import io.trino.spi.TrinoException;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
 import static io.trino.plugin.bigquery.BigQueryErrorCode.BIGQUERY_VIEW_DESTINATION_TABLE_CREATION_FAILED;
-import static io.trino.plugin.bigquery.BigQueryUtil.convertToBigQueryException;
 import static io.trino.spi.StandardErrorCode.NOT_SUPPORTED;
 import static java.lang.String.format;
-import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
 
 // A helper class, also handles view materialization
@@ -130,7 +121,7 @@ public class ReadSessionCreator
             String query = bigQueryClient.selectSql(table.getTableId(), requiredColumns);
             log.debug("query is %s", query);
             try {
-                return destinationTableCache.get(query, new DestinationTableBuilder(bigQueryClient, config, query, table.getTableId()));
+                return destinationTableCache.get(query, new BigQueryClient.DestinationTableBuilder(bigQueryClient, config, query, table.getTableId()));
             }
             catch (ExecutionException e) {
                 throw new TrinoException(BIGQUERY_VIEW_DESTINATION_TABLE_CREATION_FAILED, "Error creating destination table", e);
@@ -140,65 +131,6 @@ public class ReadSessionCreator
             // not regular table or a view
             throw new TrinoException(NOT_SUPPORTED, format("Table type '%s' of table '%s.%s' is not supported",
                     tableType, table.getTableId().getDataset(), table.getTableId().getTable()));
-        }
-    }
-
-    private static class DestinationTableBuilder
-            implements Callable<TableInfo>
-    {
-        private final BigQueryClient bigQueryClient;
-        private final ReadSessionCreatorConfig config;
-        private final String query;
-        private final TableId table;
-
-        DestinationTableBuilder(BigQueryClient bigQueryClient, ReadSessionCreatorConfig config, String query, TableId table)
-        {
-            this.bigQueryClient = requireNonNull(bigQueryClient, "bigQueryClient is null");
-            this.config = requireNonNull(config, "config is null");
-            this.query = requireNonNull(query, "query is null");
-            this.table = requireNonNull(table, "table is null");
-        }
-
-        @Override
-        public TableInfo call()
-        {
-            return createTableFromQuery();
-        }
-
-        TableInfo createTableFromQuery()
-        {
-            TableId destinationTable = bigQueryClient.createDestinationTable(table);
-            log.debug("destinationTable is %s", destinationTable);
-            JobInfo jobInfo = JobInfo.of(
-                    QueryJobConfiguration
-                            .newBuilder(query)
-                            .setDestinationTable(destinationTable)
-                            .build());
-            log.debug("running query %s", jobInfo);
-            Job job = waitForJob(bigQueryClient.create(jobInfo));
-            log.debug("job has finished. %s", job);
-            if (job.getStatus().getError() != null) {
-                throw convertToBigQueryException(job.getStatus().getError());
-            }
-            // add expiration time to the table
-            TableInfo createdTable = bigQueryClient.getTable(destinationTable);
-            long expirationTime = createdTable.getCreationTime() +
-                    TimeUnit.HOURS.toMillis(config.viewExpirationTimeInHours);
-            Table updatedTable = bigQueryClient.update(createdTable.toBuilder()
-                    .setExpirationTime(expirationTime)
-                    .build());
-            return updatedTable;
-        }
-
-        Job waitForJob(Job job)
-        {
-            try {
-                return job.waitFor();
-            }
-            catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new BigQueryException(BaseServiceException.UNKNOWN_CODE, format("Job %s has been interrupted", job.getJobId()), e);
-            }
         }
     }
 }
